Using Javascript interop to find out the Xposition of the current frame. Xposition is used to position the keyboard under the current frame.

+++ Classes/ScoreInput.cs
@@ -6,7 +6,7 @@ namespace BlazorBowlingScoreCard.Classes
 {
     public class ScoreInput : ComponentBase
     {
-        private double _xPos;
+        private int _xPos;
 
         [Parameter]
         public EventCallback<int> ScoreClicked { get; set; }
@@ -27,7 +27,7 @@ namespace BlazorBowlingScoreCard.Classes
         public bool IsVisible { get; set; }
 
         [Parameter]
-        public double XPos { get => _xPos; set => _xPos = value - 200; }
+        public int XPos { get => _xPos; set => _xPos = value - 200; }
 
         public bool Disabled(int buttonNumber)
         {
